# lib/api-zod/src/generated/types/createCheckinBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaathiCare API - Companion app for Indian senior citizens
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCheckinBodyEnergy } from "./createCheckinBodyEnergy";
import type { CreateCheckinBodyMood } from "./createCheckinBodyMood";

export interface CreateCheckinBody {
  mood: CreateCheckinBodyMood;
  energy: CreateCheckinBodyEnergy;
  notes?: string;
}