# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaathiCare API - Companion app for Indian senior citizens
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactBodyRelationship } from "./createContactBodyRelationship";

export interface CreateContactBody {
  name: string;
  relationship: CreateContactBodyRelationship;
  phone: string;
  isEmergency: boolean;
  avatar?: string;
}