# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaathiCare API - Companion app for Indian senior citizens
 * OpenAPI spec version: 0.1.0
 */
import type { Reminder } from "./reminder";

export interface DashboardSummary {
  userName: string;
  todayRemindersTotal: number;
  todayRemindersCompleted: number;
  checkinStreak: number;
  todayCheckinDone: boolean;
  activeMedicationsCount: number;
  latestMood?: string;
  weeklyHealthRecords: number;
  upcomingReminders: Reminder[];
  motivationalMessage?: string;
}